Continue some cleanup. Reduce number of modules in ui
